Sweeping refactor and test changes to Atomizer
Covers: GH-33, GH-30, GH-27, GH-22

Includes separation of conductors to their own repository, as well as
the move to the devnw GitHub organization. This changed the module name
as well as the path for pulling the module.

These changes include new event and error types which are exported. The
Errors method has been removed in lieu of the Events() method which
returns a slice of interface{} and all events and errors flow out of
Atomizer through that method when the user cares for those
events/errors.

A number of deadlocks and race conditions have been corrected throughout
the code and unit tests, and the code has been re-formatted to resemble
more idiomatic Go formatting. The validator library has been upgraded
and Atomizer has been configured to use the new Valid method as well as
Assert where applicable.