Adding constructor registration, and state copying
The `Atomize` method now accepts registration data for adding Atoms and
Conductors through the constructor. Along with this change the Register
method in the package now allows for variadic arguments when
registering.

A new feature that was added in this changeset is the ability to copy
state from an Atom registration when executing it with a payload. This
is triggered through the "CopyState" option in the electron. CopyState
only works on exported struct fields in the Atom and WILL NOT copy
unexported fields.

This change also updates all tests to use Fatal and Fatalf to exit on
failure earlier.

Also, some of the tests have been updated to use the new testing cleanup
methods which were made available in 1.15.